# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Qapac API - Transporte Publico de Cajamarca
  description: |
    API REST para la aplicacion movil de transporte publico en Cajamarca, Peru.

    ## Audiencia
    Equipo de desarrollo movil (Android) que consume esta API.

    ## Autenticacion
    - Los endpoints **publicos** no requieren autenticacion (usuarios anonimos).
    - Los endpoints de **conductor** y **admin** requieren un JWT en el header `Authorization: Bearer <token>`.
    - Los tokens se obtienen via `POST /api/v1/auth/login`.
    - El access token tiene un TTL de 15 minutos; el refresh token 7 dias.
    - Se usa rotacion de refresh tokens: cada refresh emite un nuevo par.

    ## Formato de errores
    Todos los errores retornan:
    ```json
    {"error": "<mensaje>"}
    ```

    ## Polling de posiciones en tiempo real
    No hay WebSocket. El cliente movil debe hacer polling periodico a los endpoints
    de posicion de vehiculos (ej. cada 5-10 segundos).
  version: 1.0.0
  contact:
    name: Equipo Backend Qapac
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
  - url: https://api.qapac.pe
    description: Servidor de produccion (futuro)
...
paths:
...
  /api/v1/uploads/images/{filename}:
    get:
      tags: [Subida de archivos]
      summary: Descargar una imagen subida
      description: |
        Sirve una imagen previamente subida. Retorna el archivo directamente
        con su Content-Type correspondiente.
      operationId: serveImage
      parameters:
        - name: filename
          in: path
          required: true
          description: Nombre del archivo (UUID generado al subir)
          schema:
            type: string
            example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890.jpg"
      responses:
        "200":
          description: Imagen servida
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Archivo no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "file not found"

  # ===========================================================================
  # Conductor (requiere JWT con role driver o admin)
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido via POST /api/v1/auth/login.
        Header: `Authorization: Bearer <access_token>`
...
  responses:
    BadRequest:
      description: Solicitud mal formada o parametros invalidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Mensaje descriptivo del error

    # --- Auth ---
...